+++ lib/generators/cms/widget/google_maps/templates/app/concerns/cms/attributes/google_maps_map_type.rb
@@ -1,0 +1,19 @@
+module Cms
+  module Attributes
+    module GoogleMapsMapType
+      def google_maps_map_type_attribute
+        :google_maps_map_type
+      end
+
+      def default_google_maps_map_type
+        'ROADMAP'
+      ensure
+        Cms::Errors::MissingAttribute.notify(self, __FILE__)
+      end
+
+      def google_maps_map_type
+        self[google_maps_map_type_attribute] || default_google_maps_map_type
+      end
+    end
+  end
+end