+++ lib/generators/cms/widget/google_maps/templates/app/concerns/cms/attributes/latitude.rb
@@ -1,0 +1,23 @@
+module Cms
+  module Attributes
+    module Latitude
+      def latitude_attribute
+        :latitude
+      end
+
+      def default_latitude
+        0.0
+      ensure
+        Cms::Errors::MissingAttribute.notify(self, __FILE__)
+      end
+
+      def latitude
+        if self[latitude_attribute].present?
+          self[latitude_attribute].to_f
+        else
+          default_latitude
+        end
+      end
+    end
+  end
+end