+++ lib/generators/cms/widget/google_maps/templates/app/models/google_maps_pin.rb
@@ -1,0 +1,12 @@
+class GoogleMapsPin < Obj
+  include Cms::Attributes::GoogleMapsAddress
+
+  def as_json(options = {})
+    {
+      :title => self.title,
+      :body => self.body,
+      :latitude => self.latitude,
+      :longitude => self.longitude,
+    }
+  end
+end