+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 = v0.0.5
+  * A new google maps box type was added. It allows to easily create a map and place pins on it.
+    The new box type is part of the base Infopark Kickstarter functionality but is also
+    available as a separate generator. Call ```rails generate cms:widget:google_maps --help```
+    for more information.
   * Added three additional rake tasks to easily edit the cloud config file, that holds configuration
     parameters that should not be checked in to the version control system. See
     ```rake -T cms:cloud_config``` for more details on the tasks.